Polish R output Polish progress shown in status bar Lines starting

+++ src/plugins/boxes/simulation.h
@@ -28,7 +28,7 @@ private:
     bool hasError;
     QString errorMsg;
     // Data
-    int nextShowProgress;
+    double nextShowProgress;
     QList<base::Port*> _trackedPorts;
     QFile _file;
     QTextStream _stream;
